pyproject.toml
[project]
name = "project-magpie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.9",
    "pymongo>=4.10.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "pyupbit>=0.2.34",
    "redis>=7.2.0",
    "ta-lib>=0.6.8",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (ruff format이 처리)
    "B008",  # function call in default argument
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
ignore_missing_imports = true
